Extract issues_navigation module to keep mod.rs under hard limit
